refact: improve component organization in experience section
