+++ src/pages/ExperienceSection/components/RightSection/components/Education/consts.ts
@@ -1,0 +1,87 @@
+import type { ExperienceItem } from '../../types'
+
+export const EDUCATION_ITEMS: ExperienceItem[] = [
+  {
+    date: {
+      en: 'Dec 2025',
+      pt: 'Dez 2025',
+    },
+    title: {
+      en: 'Systems Analysis and Development',
+      pt: 'Análise e Desenvolvimento de Sistemas',
+    },
+    company: 'Fatec Ferraz de Vasconcelos',
+  },
+  {
+    date: {
+      en: 'Oct 2024',
+      pt: 'Out 2024',
+    },
+    title: {
+      en: 'Full Stack JavaScript',
+      pt: 'Formação Full stack JavaScript',
+    },
+    company: 'Alura',
+    link: 'https://cursos.alura.com.br/degree/certificate/fb3108ca-577d-4444-9a49-88d6a9b5334d?lang',
+  },
+  {
+    date: {
+      en: 'Sep 2024',
+      pt: 'Set 2024',
+    },
+    title: {
+      en: 'Authentication, Testing and Security in Node.js',
+      pt: 'Formação Autenticação, testes e segurança em Node.js',
+    },
+    company: 'Alura',
+    link: 'https://cursos.alura.com.br/degree/certificate/9f510582-f1fe-478e-a72f-fe0a631bbf04?lang',
+  },
+  {
+    date: {
+      en: 'May 2024',
+      pt: 'Mai 2024',
+    },
+    title: {
+      en: 'Java Web: Create Applications Using Spring Boot',
+      pt: 'Formação Java Web: crie aplicações usando Spring Boot',
+    },
+    company: 'Alura',
+    link: 'https://cursos.alura.com.br/degree/certificate/56239697-ece2-40d5-870d-ed852b250483?lang',
+  },
+  {
+    date: {
+      en: 'Apr 2024',
+      pt: 'Abr 2024',
+    },
+    title: {
+      en: 'APIs with Node.js and Express',
+      pt: 'Formação APIs com Node.js e Express',
+    },
+    company: 'Alura',
+    link: 'https://cursos.alura.com.br/degree/certificate/61798eaf-00f8-4ea5-a417-22405b15ad04?lang',
+  },
+  {
+    date: {
+      en: 'Nov 2023',
+      pt: 'Nov 2023',
+    },
+    title: {
+      en: 'Java and Spring Boot',
+      pt: 'Formação Java e Spring Boot',
+    },
+    company: 'Alura',
+    link: 'https://cursos.alura.com.br/degree/certificate/10d216f2-a7c6-4b3a-b96f-32b7fae85217',
+  },
+  {
+    date: {
+      en: 'Dec 2022',
+      pt: 'Dez 2022',
+    },
+    title: {
+      en: 'Technician in Mechatronics',
+      pt: 'Técnico em Mecatrônica',
+    },
+    company: 'ETEC Presidente Vargas',
+    link: '',
+  },
+]
