+++ src/pages/ExperienceSection/components/RightSection/components/Experience/consts.ts
@@ -1,0 +1,105 @@
+import type { ExperienceItem } from '../../types'
+
+export const EXPERIENCE_ITEMS: ExperienceItem[] = [
+  {
+    date: {
+      en: 'Jun 2025 - Present · 8 months',
+      pt: 'Jun 2025 - Presente · 8 meses',
+    },
+    title: {
+      en: 'Junior Fullstack Developer',
+      pt: 'Desenvolvedor Fullstack Júnior',
+    },
+    company: 'Valcenter Planejados',
+    skills: [
+      'TypeScript',
+      'React.js',
+      'React Three Fiber',
+      'Full-time',
+      'Hybrid',
+    ],
+    description: [
+      {
+        en: 'Development and maintenance of back-end systems using TypeScript.',
+        pt: 'Desenvolvimento e manutenção de sistemas back-end utilizando TypeScript.',
+      },
+      {
+        en: 'Front-end development using React.js with React Three Fiber.',
+        pt: 'Desenvolvimento front-end utilizando React.js com React Three Fiber.',
+      },
+    ],
+  },
+  {
+    date: {
+      en: 'Jul 2024 - Jun 2025 · 1 year',
+      pt: 'Jul 2024 - Jun 2025 · 1 ano',
+    },
+    title: {
+      en: 'IT Intern',
+      pt: 'Estagiário de TI',
+    },
+    company: 'NewGo Tecnologia',
+    skills: [
+      'Express.js',
+      'TypeScript',
+      'React.js',
+      'EJS',
+      'Jest',
+      'Vitest',
+      'Cypress',
+      'Docker',
+      'PostgreSQL',
+      'Agile Methodology',
+    ],
+    description: [
+      {
+        en: 'Development and maintenance of back-end systems using Express.js with TypeScript.',
+        pt: 'Desenvolvimento e manutenção back-end utilizando Express.js com TypeScript.',
+      },
+      {
+        en: 'Front-end development using EJS and React.js.',
+        pt: 'Desenvolvimento front-end utilizando EJS e React.js.',
+      },
+      {
+        en: 'Writing automated tests (unit, integration, and E2E) using Jest, Vitest and Cypress.',
+        pt: 'Escrita de testes automatizados (unitários, integração e E2E) utilizando Jest, Vitest e Cypress.',
+      },
+      {
+        en: 'Using Docker for containerization and PostgreSQL databases.',
+        pt: 'Utilização de Docker para containerização e bancos de dados PostgreSQL.',
+      },
+      {
+        en: 'Experience in agile delivery methodology, focusing on open communication with the team and timely delivery to clients.',
+        pt: 'Experiência em metodologia de entrega ágil, focando em comunicação aberta com a equipe e entrega para o cliente nos tempos estimados.',
+      },
+    ],
+  },
+  {
+    date: {
+      en: 'Jul 2024 - Dec 2024 · 6 months',
+      pt: 'Jul 2024 - Dez 2024 · 6 meses',
+    },
+    title: {
+      en: 'Programming and Robotics Teacher',
+      pt: 'Professor de Programação e Robótica',
+    },
+    company: 'Ctrl+Play',
+    skills: [
+      'Programming',
+      'Robotics',
+      'Embedded Systems',
+      'Teaching',
+      'Multiple Programming Languages',
+    ],
+    description: [
+      {
+        en: 'Teaching programming and robotics to students of various age groups.',
+        pt: 'Ensino de programação e robótica para alunos de diversas faixas etárias.',
+      },
+      {
+        en: 'In-depth instruction on programming in multiple languages and robotics with embedded systems.',
+        pt: 'Instrução aprofundada sobre programação em várias linguagens e robótica com sistemas embarcados.',
+      },
+    ],
+  },
+]
